Create FindRotation2.m
Theoretical forward transformation (applied by nexus) to rotate the force plate.
The transformation still needs to be tested against trails performed in Nexus.

Ones the function of the forward transformation is found we can move on to step 2.

Step 2  We need to find a set of input parameters (i.e. 3 Rotation parameters)  that result in the physical observed case. Fitting (i.e. where the observations match the measured data) input parameters are found through a "fmincon" optimization problem.